Warning cleanup
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23673 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/plots/Spreadsheet/SpreadsheetViewerPluginInfo.C
@@ -563,8 +563,6 @@ void
 SpreadsheetViewerPluginInfo::PrivateSetPlotAtts(AttributeSubject *atts, 
     const ViewerPlot *plot)
 {
-    //const char *mName = "SpreadsheetViewerPluginInfo::PrivateSetPlotAtts: ";
-
     // Search for a SIL_DOMAIN category under the current silTopSet and add
     // sets that are used.
     SpreadsheetAttributes *sAtts = (SpreadsheetAttributes *)atts;
@@ -582,7 +580,7 @@ SpreadsheetViewerPluginInfo::PrivateSetPlotAtts(AttributeSubject *atts,
         if(*collection != NULL && collection->GetRole() == SIL_DOMAIN) 
         {
             nSets = collection->GetNumberOfSubsets(); 
-            for(size_t si = 0; si < (size_t)nSets && !validName; ++si)
+            for(int si = 0; si < nSets && !validName; ++si)
             {
                 if(!firstNameSet)
                 {
