+++ src/plots/Subset/SubsetViewerPluginInfo.C
@@ -512,12 +512,12 @@ SubsetViewerPluginInfo::PrivateSetPlotAtts(AttributeSubject *atts,
           else
           {
               int origin = mesh->groupOrigin;
-              size_t nGroups = mesh->groupIdsBasedOnRange.size()-1;
-              for (size_t i = 0; i < nGroups; i++)
+              int nGroups = (int)mesh->groupIdsBasedOnRange.size()-1;
+              for (int i = 0; i < nGroups; i++)
               {
                   groupSet.insert(origin+i);
                   gIDS.push_back(origin+i);
-                  sprintf(temp, "%ld", origin+i);
+                  sprintf(temp, "%d", origin+i);
                   sv.push_back(temp);
               }
           }
